Se limpió el código y es apto ahora para correr en produccion

+++ routes/productos.js
@@ -12,6 +12,8 @@ const contenedor = new Contenedor('./files/productos', [
 
 const router = Router()
 
+const admin = false
+
 /* ------------------------ Product Endpoints ------------------------ */
 
 // GET api/productos
@@ -31,40 +33,61 @@ router.get('/:id?', async (req, res) => {
 
 // POST api/productos
 router.post('/', async (req, res, next) => {
-  const { body } = req
+  if (admin) {
+    const { body } = req
 
-  body.timestamp = Date.now()
+    body.timestamp = Date.now()
 
-  const newProductId = await contenedor.save(body)
+    const newProductId = await contenedor.save(body)
 
-  newProductId
-    ? res
-        .status(200)
-        .json({ success: 'product added with ID: ' + newProductId })
-    : res
-        .status(400)
-        .json({ error: 'invalid key. Please verify the body content' })
+    newProductId
+      ? res
+          .status(200)
+          .json({ success: 'product added with ID: ' + newProductId })
+      : res
+          .status(400)
+          .json({ error: 'invalid key. Please verify the body content' })
+  } else {
+    res.json({
+      error: -1,
+      descripcion: 'ruta api/productos metodo post no autorizada',
+    })
+  }
 })
 
 // PUT api/productos/:id
 router.put('/:id', async (req, res, next) => {
-  const { id } = req.params
-  const { body } = req
-  const wasUpdated = await contenedor.updateById(id, body)
+  if (admin) {
+    const { id } = req.params
+    const { body } = req
+    const wasUpdated = await contenedor.updateById(id, body)
 
-  wasUpdated
-    ? res.status(200).json({ success: 'product updated' })
-    : res.status(404).json({ error: 'product not found' })
+    wasUpdated
+      ? res.status(200).json({ success: 'product updated' })
+      : res.status(404).json({ error: 'product not found' })
+  } else {
+    res.json({
+      error: -1,
+      descripcion: 'ruta api/productos metodo post no autorizada',
+    })
+  }
 })
 
 // DELETE /api/productos/:id
 router.delete('/:id', async (req, res, next) => {
-  const { id } = req.params
-  const wasDeleted = await contenedor.deleteById(id)
+  if (admin) {
+    const { id } = req.params
+    const wasDeleted = await contenedor.deleteById(id)
 
-  wasDeleted
-    ? res.status(200).json({ success: 'product successfully removed' })
-    : res.status(404).json({ error: 'product not found' })
+    wasDeleted
+      ? res.status(200).json({ success: 'product successfully removed' })
+      : res.status(404).json({ error: 'product not found' })
+  } else {
+    res.json({
+      error: -1,
+      descripcion: 'ruta api/productos metodo post no autorizada',
+    })
+  }
 })
 
 module.exports = router
